Styled TaskForm
